[TASK #161] Separated TopDownCharacter into CharacterBase and PlayerCharacter
- Made CharacterBase with standard character setup for the game.
- Renamed TP_TopDownCharacter to PlayerCharacter and moved to the Characters folder.
- Renamed BP_TopDownCharacter to BP_PlayerCharacter
- Updated references to refer to CharacterBase as a generic character class for the game.
