+++ CloudboundGames/Config/DefaultEngine.ini
@@ -251,3 +251,6 @@ ManualIPAddress=
 CategorySlot2=NumPadTwo
 CategorySlot3=NumPadThree
 
+
+[CoreRedirects]
++ClassRedirects=(OldName="/Script/CloudboundGames.TP_TopDownCharacter",NewName="/Script/CloudboundGames.PlayerCharacter")
