+++ CloudboundGames/Source/CloudboundGames/TP_TopDown/TP_TopDownPlayerController.cpp
@@ -2,7 +2,7 @@
 
 #include "TP_TopDownPlayerController.h"
 #include "GameFramework/Pawn.h"
-#include "TP_TopDownCharacter.h"
+#include "../Characters/PlayerCharacter.h"
 #include "Engine/World.h"
 #include "EnhancedInputSubsystems.h"
 #include "Blueprint/UserWidget.h"
@@ -35,7 +35,7 @@ void ATP_TopDownPlayerController::OnPossess(APawn* InPawn)
 	}
 
 	ACharacter* testPawn = GetCharacter();
-	ATP_TopDownCharacter* TopDownCharacter = Cast<ATP_TopDownCharacter, ACharacter>(testPawn);
+	APlayerCharacter* TopDownCharacter = Cast<APlayerCharacter, ACharacter>(testPawn);
 	if (!TopDownCharacter)
 	{
 		UE_LOG(LogTemplateCharacter, Error, TEXT("'%s' Failed to find a Character!"), *GetNameSafe(this));
@@ -43,12 +43,12 @@ void ATP_TopDownPlayerController::OnPossess(APawn* InPawn)
 	}
 		
 	// Setup movement events.
-	EnhancedInputComponent->BindAction(HorizontalMovementAction, ETriggerEvent::Triggered, TopDownCharacter, &ATP_TopDownCharacter::OnHorizontalMovementTriggered);
+	EnhancedInputComponent->BindAction(HorizontalMovementAction, ETriggerEvent::Triggered, TopDownCharacter, &APlayerCharacter::OnHorizontalMovementTriggered);
 	TopDownCharacter->MoveHorizontalActionBinding = &EnhancedInputComponent->BindActionValue(HorizontalMovementAction);
-	EnhancedInputComponent->BindAction(VerticalMovementAction, ETriggerEvent::Triggered, TopDownCharacter, &ATP_TopDownCharacter::OnVerticalMovementTriggered);
+	EnhancedInputComponent->BindAction(VerticalMovementAction, ETriggerEvent::Triggered, TopDownCharacter, &APlayerCharacter::OnVerticalMovementTriggered);
 	TopDownCharacter->MoveVerticalActionBinding = &EnhancedInputComponent->BindActionValue(VerticalMovementAction);
 
-	EnhancedInputComponent->BindAction(RollAction, ETriggerEvent::Triggered, TopDownCharacter, &ATP_TopDownCharacter::OnRollTriggered);
+	EnhancedInputComponent->BindAction(RollAction, ETriggerEvent::Triggered, TopDownCharacter, &APlayerCharacter::OnRollTriggered);
 
 	HUDWidget = CreateWidget<UHUDWidget_Base, ATP_TopDownPlayerController*>(this, HUDWidgetType, "HUD");
